+++ docs/design/src/interface-adapters/mappers/RewriteRuleMapper/getAllRules.md
@@ -48,10 +48,10 @@ tests/unit/interface-adapters/mappers/RewriteRuleMapper/getAllRules/
 
 ### モック方法
 
-共通モックファクトリからインポートし、`beforeEach`で初期化（コンストラクタインジェクション）:
+インターフェース配下のモックファクトリからインポートし、`beforeEach`で初期化（コンストラクタインジェクション）:
 
 ```typescript
-import { createMockRewriteRuleMessagingPort } from 'tests/unit/shared/mocks/createMockRewriteRuleMessagingPort';
+import { createMockRewriteRuleMessagingPort } from 'tests/unit/interface-adapters/ports/IRewriteRuleMessagingPort/mocks/createMockRewriteRuleMessagingPort';
 
 let mockMessagingPort: IRewriteRuleMessagingPort;
 
@@ -71,9 +71,9 @@ const mapper = new RewriteRuleMapper(mockMessagingPort);
 
 ### モックファイル構成
 
-共通モックファクトリを使用（複数テストで共有）:
+インターフェース配下にモックファクトリを配置（複数テストで共有）:
 
 ```
-tests/unit/shared/mocks/
-└── createMockRewriteRuleMessagingPort.ts    # 共通モックファクトリ
+tests/unit/interface-adapters/ports/IRewriteRuleMessagingPort/mocks/
+└── createMockRewriteRuleMessagingPort.ts    # モックファクトリ
 ```
