+++ docs/design/src/interface-adapters/mappers/RewriteRuleMapper/toEntity.md
@@ -44,10 +44,10 @@ tests/unit/interface-adapters/mappers/RewriteRuleMapper/toEntity/
 
 ### モック方法
 
-共通モックファクトリからインポートし、インスタンス生成時に注入:
+インターフェース配下のモックファクトリからインポートし、インスタンス生成時に注入:
 
 ```typescript
-import { createMockRewriteRuleMessagingPort } from 'tests/unit/shared/mocks/createMockRewriteRuleMessagingPort';
+import { createMockRewriteRuleMessagingPort } from 'tests/unit/interface-adapters/ports/IRewriteRuleMessagingPort/mocks/createMockRewriteRuleMessagingPort';
 
 const mockMessagingPort = createMockRewriteRuleMessagingPort();
 const mapper = new RewriteRuleMapper(mockMessagingPort);
@@ -55,9 +55,9 @@ const mapper = new RewriteRuleMapper(mockMessagingPort);
 
 ### モックファイル構成
 
-共通モックファクトリを使用（複数テストで共有）:
+インターフェース配下にモックファクトリを配置（複数テストで共有）:
 
 ```
-tests/unit/shared/mocks/
-└── createMockRewriteRuleMessagingPort.ts    # 共通モックファクトリ
+tests/unit/interface-adapters/ports/IRewriteRuleMessagingPort/mocks/
+└── createMockRewriteRuleMessagingPort.ts    # モックファクトリ
 ```
