+++ host-frontend-root/frontend-src-root/tests/unit/frameworks-and-drivers/persistence/ChromeRuntimeRewriteRuleRepository/getRulesMatchingUrl/normal-cases.test.ts
@@ -6,7 +6,7 @@ import { ChromeRuntimeRewriteRuleRepository } from 'src/frameworks-and-drivers/p
 import { RewriteRuleMapper } from 'src/interface-adapters/mappers/RewriteRuleMapper';
 import { IRewriteRuleMessagingPort } from 'src/interface-adapters/ports/IRewriteRuleMessagingPort';
 
-import { createMockRewriteRuleMessagingPort } from 'tests/unit/shared/mocks/createMockRewriteRuleMessagingPort';
+import { createMockRewriteRuleMessagingPort } from 'tests/unit/interface-adapters/ports/IRewriteRuleMessagingPort/mocks/createMockRewriteRuleMessagingPort';
 
 /**
  * ChromeRuntimeRewriteRuleRepository.getRulesMatchingUrl - 正常系テスト
