+++ host-frontend-root/frontend-src-root/tests/unit/interface-adapters/ports/IRewriteRuleMessagingPort/mocks/createMockRewriteRuleMessagingPort.ts
